+++ test/ConvergenceTests/runtests.jl
@@ -14,8 +14,9 @@ using Test
   @time @testset "Elasticity - HHO (mixed order)" begin include("HHOMixedElasticity.jl") end
 end
 
-@testset "C1 elements" begin
+@testset "Fourth-order plate elements" begin
   @time @testset "Biharmonic - Argyris" begin include("Argyris.jl") end
+  @time @testset "Biharmonic - Morley" begin include("Morley.jl") end
 end
 
 end